implement named pipes for remoteapi on windows
