+++ monitor/remoteapi/client/client.go
@@ -13,20 +13,7 @@ import (
 	"go.aporeto.io/trireme-lib/common"
 )
 
-// Client is an api client structure.
-type Client struct {
-	addr *net.UnixAddr
-}
-
-// NewClient creates a new client.
-func NewClient(path string) (*Client, error) {
-	addr, err := net.ResolveUnixAddr("unix", path)
-	if err != nil {
-		return nil, fmt.Errorf("invalid address: %s", err)
-	}
-
-	return &Client{addr: addr}, nil
-}
+type dialContextFunc func(ctx context.Context, network, address string) (net.Conn, error)
 
 // SendRequest sends a request to the remote.
 // TODO: Add retries
@@ -34,9 +21,7 @@ func (c *Client) SendRequest(event *common.EventInfo) error {
 
 	httpc := http.Client{
 		Transport: &http.Transport{
-			DialContext: func(_ context.Context, _, _ string) (net.Conn, error) {
-				return net.DialUnix("unix", nil, c.addr)
-			},
+			DialContext:     c.getDialContext(),
 			MaxIdleConns:    10,
 			IdleConnTimeout: 10 * time.Second,
 		},
